Added file to simplify format string creation
This file will contain any macros and other such specific implementation details that will be related to format strings so as to make it easier to write them.

We could just use raw strings and save a lot of trouble, but I'm trying to keep things somewhat C89 / ANSI C compliant for crossplatformness reasons.