Fixed errors in main program
The generator could only work with up to as many usernames as were stored within the usernames file. This limitation has been temporarily fixed by using a modulo operation. The usernames are repeated, which is not ideal, but it is better than the program not working whenever the user asks for an input with more than 100 users within a list. The ideal fix would be to use a buffer and snprintf or some other such method to generate a few numbers after the username if username duplication is not desired, but since that does not matter for this project (all that matters is unique IDs), then this is a good enough fix.

The OID string for the second commenter was chosen wrongly due to a typo in the code, leading to UB during file generation (in the case of the systems where this was tested, an empty string would be printed to the file, but that is not guaranteed since we were attempting to print a pointer to God knows where in memory... it was silently failing because coincidentally it was always pointing to a segment with null values, which prevented the tests from generating garbage or even crashing). That error has now been fixed, and the correct string is now properly chosen.