Modified program to include extra options
The program now allows the user to select wether they want to store all entries within a single file or each entry within an individual file each.

This was done for ease of importing in different software. For our specific usecase, having all entries within a single file makes it easier to batch import in MongoDB.